refactor(services): share install result presentation
