+++ lib/hive/commands/bot.rb
@@ -10,10 +10,13 @@ require "hive/bot/logger"
 require "hive/pid_file"
 require "hive/paths"
 require "hive/update_check/state"
+require "hive/commands/service_installer/result_presenter"
 
 module Hive
   module Commands
     class Bot
+      include Hive::Commands::ServiceInstaller::ResultPresenter
+
       VALID_SUBCOMMANDS = %w[start stop status reload tail install].freeze
       JSON_USAGE_ERROR_SCHEMA = "hive-bot-status".freeze
 
@@ -260,141 +263,17 @@ module Hive
       def install_bot
         require "hive/commands/bot/service_installer"
         installer = Hive::Commands::Bot::ServiceInstaller.new(binary_path: current_binary_path)
-        begin
-          result = installer.install!(autostart: true, force: @force)
-        rescue Hive::Error
-          raise
-        rescue StandardError => e
-          install_emit_exception_envelope(installer, e) if @json
-          raise Hive::BotInstallFailed,
-                "bot service install failed: #{e.class}: #{e.message}"
-        end
-        unless @json
-          installer.messages.each { |line| warn "hive: #{line}" }
-          emit_install_success_summary(installer, result)
-        end
-        emit_install_outcome(installer, result)
+        perform_service_install(installer)
       end
 
-      # Bare-text positive confirmation on the non-JSON success path so
-      # operators can distinguish first-time install / no-op / in-place
-      # upgrade at a glance. Mirrors `daemon#emit_install_success_summary`.
-      def emit_install_success_summary(installer, outcome)
-        return if @json
-
-        case outcome.kind
-        when :written
-          puts "hive bot: installed unit at #{installer.target_path}"
-        when :upgraded
-          msg = "hive bot: upgraded unit at #{installer.target_path}"
-          msg += " (backup: #{outcome.backup_path})" if outcome.backup_path
-          puts msg
-        when :unchanged
-          puts "hive bot: unit already up to date at #{installer.target_path}"
-        when :autostart_unavailable
-          puts "hive bot: unit written at #{installer.target_path}; autostart not enabled on this host"
-        when :unsupported, :drifted, :failed
-          # :unsupported is messaged via installer.messages.
-          # :drifted / :failed are handled by emit_install_outcome
-          # (which raises); no positive summary applies.
-        end
-      end
-
-      def emit_install_outcome(installer, outcome)
-        if @json
-          if outcome.success?
-            puts JSON.generate(install_envelope(installer, outcome))
-          else
-            install_emit_error_envelope(installer, outcome: outcome.wire_outcome)
-          end
-        end
-
-        if outcome.drifted?
-          msg = "bot unit at #{installer.target_path} differs from the current template. " \
-                "Re-run with `hive bot install --force` to overwrite (a timestamped .bak " \
-                "will be saved)."
-          raise Hive::BotInstallDriftError, msg
-        elsif outcome.failed?
-          raise Hive::BotInstallFailed,
-                "bot service install reported a failure; see messages above"
-        end
-      end
-
-      def install_envelope(installer, outcome)
-        {
-          "schema" => "hive-bot-install",
-          "schema_version" => Hive::Schemas::SCHEMA_VERSIONS.fetch("hive-bot-install"),
-          "ok" => true,
-          "outcome" => outcome.wire_outcome,
-          "platform" => installer.envelope_platform,
-          "target_path" => installer.target_path,
-          "backup_path" => outcome.backup_path,
-          "restarted" => outcome.restarted,
-          "messages" => installer.messages.dup
-        }
-      end
-
-      def install_emit_error_envelope(installer, outcome:)
-        error_class = outcome == "drifted" ? "BotInstallDriftError" : "BotInstallFailed"
-        exit_code = outcome == "drifted" ? Hive::ExitCodes::USAGE : Hive::ExitCodes::SOFTWARE
-        message =
-          if outcome == "drifted"
-            "bot unit at #{installer.target_path} differs from the current template; retry with --force."
-          else
-            "bot service install reported a failure; see messages"
-          end
-        puts JSON.generate(
-          "schema" => "hive-bot-install",
-          "schema_version" => Hive::Schemas::SCHEMA_VERSIONS.fetch("hive-bot-install"),
-          "ok" => false,
-          "error_class" => error_class,
-          "error_kind" => outcome,
-          "exit_code" => exit_code,
-          "message" => message,
-          "outcome" => outcome,
-          "platform" => installer.envelope_platform,
-          "target_path" => installer.target_path,
-          "messages" => installer.messages.dup
-        )
-      end
-
-      def install_emit_exception_envelope(installer, error)
-        puts JSON.generate(
-          "schema" => "hive-bot-install",
-          "schema_version" => Hive::Schemas::SCHEMA_VERSIONS.fetch("hive-bot-install"),
-          "ok" => false,
-          "error_class" => "BotInstallFailed",
-          "error_kind" => "failed",
-          "exit_code" => Hive::ExitCodes::SOFTWARE,
-          "message" => "bot service install failed: #{error.class}: #{error.message}",
-          "outcome" => "failed",
-          "platform" => safe_install_platform(installer),
-          "target_path" => safe_install_target_path(installer),
-          "messages" => safe_install_messages(installer)
-        )
-      end
+      def service_install_label = "bot"
+      def service_install_schema = "hive-bot-install"
+      def service_install_drift_error = Hive::BotInstallDriftError
+      def service_install_failure_error = Hive::BotInstallFailed
 
       def current_binary_path
         require "hive/invoked_binary"
         Hive::InvokedBinary.path
-      end
-
-      def safe_install_platform(installer)
-        installer.envelope_platform
-      rescue StandardError
-        "unsupported"
-      end
-
-      def safe_install_target_path(installer)
-        installer.target_path
-      rescue StandardError
-        nil
-      end
-
-      def safe_install_messages(installer)
-        installer.messages.dup
-      rescue StandardError
-        []
       end
 
       def live_pid
