+++ wiki/testing.md
@@ -85,7 +85,8 @@ task default: :test
 | `commands/digest_test.rb` | `Hive::Commands::Digest` — strict `YYYY-MM-DD` parsing, default runner invocation, dry-run message output, `hive-digest` JSON/error envelopes, merged-PR source dispatch, `--repo` implying merged-PR mode, merged-PR human wording, and `hive-merged-pr-digest` schema validation for dry-run and real-send payloads. |
 | `commands/daemon_test.rb` | `Hive::Commands::Daemon` and `Hive::Daemon::StatusReport` — lifecycle command routing, PID-file ownership/status handling, detached start/re-exec behavior, service install/enable/disable output, daemon-status service/binary fields, `safe_payload` web degradation, `binary_drift` classification including `unreadable`, queue inspection, start-path wiring of daemon/update/digest config into the dispatcher, and `daemon.auto_retry.enabled` config threading into the dispatcher. Start tests stub all three global config blocks so operator-local Telegram digest defaults cannot change unit expectations. |
 | `commands/setup/*_test.rb` | `Hive::Commands::Setup` — phase-success predicate and JSON/process-exit agreement, including the shared `phase(name)` StandardError-to-`ok:false` wrapper for QMD bootstrap, web-bundle, daemon-service, enrollment, and optional web-service phases, plus hard diagnostic failures. |
-| `commands/daemon/service_installer_test.rb`, `commands/service_installer/base_test.rb` | Shared service installer behavior — systemd/launchd rendering, force-upgrade backups/restarts, unsupported-platform service-state probes, Homebrew stable-binary selection, macOS launchd ProgramArguments `$0` parsing for the daemon binary, and shared launchd path substitution used by daemon/web plists. |
+| `commands/{bot,daemon}/service_installer_test.rb`, `commands/service_installer/base_test.rb` | Shared service installer behavior — systemd/launchd rendering, force-upgrade backups/restarts, unsupported-platform service-state probes, Homebrew stable-binary selection, macOS launchd ProgramArguments `$0` parsing for the daemon binary, and shared launchd path substitution used by daemon/web plists. |
+| `commands/{bot,daemon}_test.rb`, `integration/daemon_command_test.rb`, `integration/bot/bot_lifecycle_test.rb` | Shared service-install result presentation — bot/daemon-specific text prefixes and schemas, every success/outcome mapping, drift exit 64, manager failure exit 70, force/backup guidance, hostile-installer fallback envelopes, and schema-valid subprocess behavior after both commands converge on `ServiceInstaller::ResultPresenter`. |
 | `digest/window_test.rb`, `digest/ship_times_test.rb`, `digest/collector_test.rb` | Digest collection primitives — local-date helpers, git-log ship-time preference (`pr_finalized`, `archived`, approval into `9-done`), registered-project grouping, missing artifact tolerance, and local timezone boundaries. |
 | `digest/categorizer_test.rb`, `digest/renderer_test.rb`, `digest/run_test.rb`, `digest/sender_test.rb`, `digest/merged_pr/*_test.rb` | Digest generation/delivery — model JSON mapping and fallbacks, prompt rendering with PR bodies, Telegram MarkdownV2 escaping/category ordering, empty/success/failed-notice orchestration, dry-run token bypass, chat-id resolution, Telegram send arguments through an injected client, merged-PR repo resolution, local-date filtering, partial GitHub failure tolerance, Hive task annotation, mechanical rendering, and runner orchestration. These are unit seams; no real agent, Telegram Bot API call, or live GitHub merged-PR query is exercised. |
 | `daemon/digest_scheduler_test.rb` | `Hive::Daemon::DigestScheduler` — first-run no-history guard, local-midnight due calculation, one-day-at-a-time catch-up, catch-up cap logging, non-zero retry behavior, disabled mode, and DST local-date handling. |
